read QR/requirements.txt
pyzbar==0.1.8
Pillow==8.4.0
